@@ -2,7 +2,7 @@
 name = "laboneq-applications-documentation"
 version = "0.1.0"
 description = "LabOne Q Applications Documentation"
-requires-python = ">=3.9"
+requires-python = ">=3.10"
 dependencies = ["mkdocs-zhinst>=0.0.26"]
 
 [tool.uv.sources]
