@@ -7,7 +7,7 @@ name = "laboneq_applications"
 description = "Library of LabOne Q experiments, analysis functions, and more."
 version = "2.7.0dev0"
 license = { text = "Apache 2.0" }
-requires-python = ">=3.9"
+requires-python = ">=3.10"
 readme = "README.md"
 authors = [
   { name = "Zurich Instruments Development Team", email = "info@zhinst.com" },
@@ -22,7 +22,6 @@ classifiers = [
   "Operating System :: MacOS",
   "Operating System :: Microsoft :: Windows",
   "Operating System :: POSIX :: Linux",
-  "Programming Language :: Python :: 3.9",
   "Programming Language :: Python :: 3.10",
   "Programming Language :: Python :: 3.11",
   "Programming Language :: Python :: 3.12",
